move add to its own component
