+++ app/javascript/components/WhiskyTracker.jsx
@@ -12,10 +12,10 @@ import {
 } from '@planningcenter/tapestry-react'
 import _ from 'lodash'
 import { keysToCamelCase } from "../utils/keysToSnakeCase"
-import  AddWhisky  from "./apis/whiskies/add"
-import Show from "./whisky/show"
 import DeleteWhisky from "./apis/whiskies/remove"
 import GetData from "./apis/getData"
+import Show from "./whisky/show"
+import Add from "./whisky/add"
 
 export default function HelloComponent({}) {
 
@@ -23,18 +23,9 @@ export default function HelloComponent({}) {
   const [locations, setLocations] = useState([])
   const [selectedWhisky, setSelectedWhisky] = useState({})
   const [selectedLocation, setSelectedLocation] = useState({})
-  const [open, setOpen] = useState(false)
-  const [whiskyName, setWhiskyName] = useState('')
-  const [tastingNotes, setTastingNotes] = useState('')
-  const [rating, setRating] = useState(0)
-  const [locationId, setLocationId] = useState(0)
+  const [openAdd, setOpenAdd] = useState(false)
+  const [openShow, setOpenShow] = useState(false)
   const [showWhisky, setShowWhisky] = useState(false)
-
-  const addWhisky = async () => {
-    const newWhisky = await AddWhisky({name: whiskyName, tastingNotes, rating, locationId})
-    setWhiskies((prev) => [...prev, newWhisky])
-    setOpen(false)
-  }
 
   const removeWhisky = async (whisky) => {
     await DeleteWhisky(whisky)
@@ -73,54 +64,16 @@ export default function HelloComponent({}) {
           <Button
             padding={1}
             variant='naked'
-            title={open ? 'Done' : 'Add'}
+            title={openAdd ? 'Done' : 'Add'}
             theme={'primary'}
-            onClick={() => setOpen(!open)}
+            onClick={() => setOpenAdd(!openAdd)}
           />
         </StackView>
-        <Collapse open={open} lazy>
-          <StackView spacing={1}>
-            <Input
-              placeholder={'name'}
-              onChange={e => setWhiskyName(e.target.value)}
-              autoFocus
-            />
-            <Input
-              placeholder={'rating'}
-              onChange={e => setRating(e.target.value)}
-            />
-            <Input
-              placeholder={'tasting notes'}
-              onChange={e => setTastingNotes(e.target.value)}
-            />
-            <StackView>
-              <Input.InputLabel>Location</Input.InputLabel>
-                <Select
-                  id="my-select-menu"
-                  onChange={({ value }) => setLocationId(value)
-                  }
-                >
-                  {locations.map((location) => (
-                    <Select.Option
-                      value={location.id}
-                      onChange={(value)=> setLocationId(value)}
-                    >
-                      {`${location.name}`}
-                    </Select.Option>
-                  ))}
-                </Select>
-            </StackView>
-            <StackView width={15}>
-                <Button
-                  padding={1}
-                  size='sm'
-                  variant='naked'
-                  theme={'primary'}
-                  title="add whisky"
-                  onClick={() => addWhisky()}
-                />
-            </StackView>
-            </StackView>
+        <Collapse open={openAdd} lazy>
+          <Add locations={locations} setWhiskies={setWhiskies} setOpen={setOpenAdd}/>
+        </Collapse>
+        <Collapse open={showWhisky} lazy>
+          <Show whisky={selectedWhisky} location={selectedLocation} setShowWhisky={setShowWhisky}/>
         </Collapse>
         {whiskies.map((whisky) => (
           <>
@@ -134,9 +87,6 @@ export default function HelloComponent({}) {
           </>
         ))}
       </StackView>
-      {showWhisky && (
-        <Show whisky={selectedWhisky} location={selectedLocation} setShowWhisky={setShowWhisky}/>
-      )}
     </StackView>
   )
 }
